Decimals fixed and roughly tested

+++ workoutLogging/models.py
@@ -5,13 +5,13 @@ from datetime import date
 
 # Create your models here.
 class Workout(models.Model):
-    distance                =   models.FloatField()
+    distance                =   models.DecimalField(max_digits=20, decimal_places=4)
     # next line citation [16]
     shoe                    =   models.ForeignKey('settings.Shoe', on_delete=models.SET_NULL, null=True)
     entry                   =   models.TextField(default="")
-    hours                   =   models.DecimalField(null=True)
-    minutes                 =   models.DecimalField(null=True)
-    seconds                 =   models.DecimalField(null=True)
+    hours                   =   models.DecimalField(max_digits=20, decimal_places=4, null=True)
+    minutes                 =   models.DecimalField(max_digits=20, decimal_places=4, null=True)
+    seconds                 =   models.DecimalField(max_digits=20, decimal_places=4, null=True)
     # type/subtype from Merv
     # next line citation [16]
     wtype                   =   models.ForeignKey('settings.WorkoutType', null=True, on_delete=models.SET_NULL)
@@ -20,9 +20,9 @@ class Workout(models.Model):
     mervLegacySubtype       =   models.CharField(max_length=100, null=True)
     mervLegacyPace          =   models.CharField(max_length=100, null=True)
     mervLegacyPaceUnits     =   models.CharField(max_length=50, null=True)
-    mervLegacyHeartrate     =   models.DecimalField(null=True)
+    mervLegacyHeartrate     =   models.DecimalField(max_digits=20, decimal_places=4, null=True)
     mervLegacyAddendum      =   models.CharField(max_length=100, null=True)
-    mervLegacyDistance      =   models.DecimalField(null=True)
+    mervLegacyDistance      =   models.DecimalField(max_digits=20, decimal_places=4, null=True)
     mervLegacyDistanceUnits =   models.CharField(max_length=30, null=True)
     title                   =   models.CharField(max_length=100, default="")
     modifiedDate            =   models.DateTimeField()
@@ -50,4 +50,4 @@ class Comment(models.Model):
 class Unit(models.Model):
     owner       = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
     name        = models.CharField(max_length=100)
-    distance    = models.DecimalField()
+    distance    = models.DecimalField(max_digits=20, decimal_places=4)
